Completes ex 3.6, error on POST duplicate name
This commit adds a check to the POST if the data contains a name that
is already present in the phonebook. If it does, it responds with an
error 400 specifying that the name must be unique.

The previous commit already included error checking to ensure the
request data included a name and number, which is the other part of
exercise 3.6