Completes ex. 3.8, adds POST data to morgan log
This commit adds to the server logging the body data sent with a POST
request to the end of the log. This is done by adding a custom token
to morgan and adding it to the log. This also requires replacing the
'tiny' parameter with the equivalent token string.

No additional information is logged when a the request is anything
other than POST.