Completes ex. 3.10, deploying backend to fly.io
This commit adds fly.io deployment to the backend of the phonebook.
The deployed backend responds appropriately to GET requests. It does
not handle DELETE or PUSH requests, instead returning a 301 response.
Given where this appears in the course material, and looking at
future part 3 exercises, I believe this is the expected state of the
initial backend deployment.