Completes ex. 3.12, adds command line database
This commit adds command line database functionality to the app with
a MongoDB database via mongoose. The database is accessed via node
with either 1 or 3 parameters. The first of these is the password for
the DB on mongodb.net. If called with only one parameter, the script
logs the contents of the phonebook on the database. The optional 2nd
and 3rd parameter are the name and number for a new phonebook entry,
respectively. If called with 3 parameters, the script adds the new
entry to the database on mongodb.net. If called with an incorrect
number of parameters, the script prints an appropriate error message.