Completes ex. 3.15, enables deleting entries
This commit adds functionality to handle DELETE requests to the DB
to remove entries from the phonebook app. It also updates the entry
model in entry.js, creating a JSON schema to process the entries. The
schema renames the "_id" field to "id", and removes the __v field.
This schema change enables the frontend delete buttons to continue
to function with the backend & DB.