Completes ex. 3.17, enables PUT requests to DB
This commit adds handling for DB PUT requests to update the phone
numbers in the phonebook app. This also works with the frontend, so
that when an entry with an identical name is added it will update the
number of the entry (after the user confirms that they want to update
via a window.confirm prompt).