Completes ex. 3.19 & 3.20, adds data validation
This commit adds data validation to the mongoose entry schema for the
phonebook app. This requires entry names to be at least 3 characters;
and numbers to be at least 8 characters, with all but 1 character
being a digit, the non-digit being a dash, and the dash appearing 3rd
or 4th in the number. The backend index.js is updated to handle any
validation errors, and the frontend is updated to handle errors and
display appropriate error messages.